backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
qrcode[pil]==7.4.2

# Config
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-decouple==3.8

# Task queue
celery==5.3.6
redis==5.0.4

# Rate limiting
slowapi==0.1.9

# HTTP client
httpx==0.27.0
aiohttp==3.9.5

# Email / DNS
dnspython==2.6.1
python-whois==0.9.4
checkdmarc==4.6.7

# File analysis
python-magic==0.4.27
yara-python==4.5.1

# Security scanning
bandit==1.7.8

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0

# Utilities
structlog==24.1.0
pip-audit==2.7.3
